part way to figuring out the issues. but need to go update #21
