+++ src/main/webui/src/observations/timingWindowGui.tsx
@@ -1,0 +1,8 @@
+//type to use for the DateTimePickers
+export type TimingWindowGui = {
+    startTime: Date | null,
+    endTime: Date | null,
+    note: string,
+    isAvoidConstraint: boolean,
+    key: string
+}
